+++ Cryptocurrency Management.xcworkspace/contents.xcworkspacedata
@@ -2,9 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:LineChartView.swift">
-   </FileRef>
-   <FileRef
       location = "group:Cryptocurrency Management.xcodeproj">
    </FileRef>
    <FileRef
